컨트롤러 관련 메서드 및 예외 처리 개선
- `ControllerAdapter.h`: `Connect`와 `ReceiveState`의 반환 타입 변경 및 `ReceiveQueue` 메서드 추가.
- `XInputAdapter.cpp`: `Connect`와 `ReceiveState` 구현 수정, 예외 처리 추가.
- `InputExceptions`: `ThrowIfFailed` 메서드 인자 타입 변경, 새로운 예외 클래스 추가.
- `Controller.cpp`: `Connect`와 `UpdateState` 반환 타입 변경, 예외 처리 추가, 새로운 `GetButtonQueue` 메서드 추가.
- `Controller.h`: `Button` 타입을 `enum`으로 변경, 버튼 비트 정의 수정.

+++ Source/GA6thFinal_Framework/GameEngine/Source/Engine/InputCore/Declare/Exceptions/InputExceptions.cpp
@@ -3,7 +3,7 @@
 
 namespace Input
 {
-    void ThrowIfFailed::operator()(const long result, const std::string& message) const
+    void ThrowIfFailed::operator()(const DWORD result, const std::string& message) const
     {
         if (result != ERROR_SUCCESS)
         {
